# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Flottakezelő API",
      "version": "1.0.0",
      "description": "API dokumentáció a Járműflotta-kezelő backendhez."
    },
    "servers": [
      {
        "url": "https://flotta.vbdev.hu/api/",
        "description": "VB. VPS"
      }
    ],
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
      "/api/auth/register": {
        "post": {
          "tags": ["Auth"],
          "summary": "Új felhasználó regisztrációja",
          "security": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": { "type": "string", "example": "admin_user" },
                    "password": { "type": "string", "example": "admin_secret_123" },
                    "role": { "type": "string", "example": "admin" }
                  }
                }
              }
            }
          },
          "responses": {
            "201": { "description": "Sikeres regisztráció" }
          }
        }
      },
      "/api/auth/login": {
        "post": {
          "tags": ["Auth"],
          "summary": "Felhasználó bejelentkezése",
          "security": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": { "type": "string", "example": "admin_user" },
                    "password": { "type": "string", "example": "admin_secret_123" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": { "description": "Sikeres bejelentkezés" }
          }
        }
      },
...
      "/api/drivers": {
        "get": {
          "tags": ["Drivers"],
          "summary": "Összes sofőr lekérdezése",
          "responses": { "200": { "description": "Sofőrök listája" } }
        },
        "post": {
          "tags": ["Drivers"],
          "summary": "Új sofőr hozzáadása",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "license_number": { "type": "string" },
                    "address": { "type": "string" },
                    "email": { "type": "string" },
                    "phone": { "type": "string" },
                    "name": { "type": "string" },
                    "starting_date": { "type": "string", "format": "date" },
                    "user_id": { "type": "integer" }
                  }
                }
              }
            }
          },
          "responses": { "201": { "description": "Sikeres létrehozás" } }
        }
      },
      "/api/drivers/{id}": {
        "get": {
          "tags": ["Drivers"],
          "summary": "Sofőr lekérdezése ID alapján",
          "parameters": [
            { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }
          ],
          "responses": { "200": { "description": "Sofőr adatai" } }
        },
        "put": {
          "tags": ["Drivers"],
          "summary": "Sofőr módosítása",
          "parameters": [
            { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "license_number": { "type": "string" },
                    "address": { "type": "string" },
                    "email": { "type": "string" },
                    "phone": { "type": "string" },
                    "name": { "type": "string" },
                    "starting_date": { "type": "string", "format": "date" },
                    "user_id": { "type": "integer" }
                  }
                }
              }
            }
          },
          "responses": { "200": { "description": "Sikeres módosítás" } }
        },
        "delete": {
          "tags": ["Drivers"],
          "summary": "Sofőr törlése",
          "parameters": [
            { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }
          ],
          "responses": { "200": { "description": "Sikeres törlés" } }
        }
      },
      "/api/driver-assignments": {
        "get": {
          "tags": ["Driver Assignments"],
          "summary": "Összes hozzárendelés lekérdezése",
          "responses": { "200": { "description": "Hozzárendelések" } }
        },
        "post": {
          "tags": ["Driver Assignments"],
          "summary": "Új hozzárendelés",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "assigned_from": { "type": "string", "format": "date" },
                    "assigned_to": { "type": "string", "format": "date" },
                    "vehicle_id": { "type": "integer" },
                    "driver_id": { "type": "integer" }
                  }
                }
              }
            }
          },
          "responses": { "201": { "description": "Sikeres létrehozás" } }
        }
      },
      "/api/driver-assignments/{id}": {
        "get": {
          "tags": ["Driver Assignments"],
          "summary": "Hozzárendelés lekérdezése",
          "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
          "responses": { "200": { "description": "Siker" } }
        },
        "put": {
          "tags": ["Driver Assignments"],
          "summary": "Hozzárendelés módosítás",
          "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "assigned_from": { "type": "string", "format": "date" },
                    "assigned_to": { "type": "string", "format": "date" },
                    "vehicle_id": { "type": "integer" },
                    "driver_id": { "type": "integer" }
                  }
                }
              }
            }
          },
          "responses": { "200": { "description": "Sikeres módosítás" } }
        },
        "delete": {
          "tags": ["Driver Assignments"],
          "summary": "Hozzárendelés törlése",
          "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
          "responses": { "200": { "description": "Törölve" } }
        }
     	},
	"/api/driver-assignments/driver/{driver_id}": {
		"get": {
			"tags": ["Driver Assignments"],
			"summary": "Hozzárendelés lekérdezése sofőr alapján",
			"parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
			"responses": { "200": { "description": "Siker" } }
		}
	},
...
      "/api/fuelings": {
        "get": { "tags": ["Fuelings"], "summary": "Tankolások", "responses": { "200": { "description": "Siker" } } },
        "post": {
          "tags": ["Fuelings"], "summary": "Új tankolás", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "date": { "type": "string", "format": "date" }, "amount_liters": { "type": "number" }, "price_per_liter": { "type": "number" }, "vehicle_id": { "type": "integer" } } } } } }, "responses": { "201": { "description": "Létrehozva" } }
        }
      },
      "/api/fuelings/{id}": {
        "get": { "tags": ["Fuelings"], "summary": "Egy tankolás", "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Siker" } } },
        "put": {
          "tags": ["Fuelings"], "summary": "Módosítás", "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }],
          "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "date": { "type": "string", "format": "date" }, "amount_liters": { "type": "number" }, "price_per_liter": { "type": "number" }, "vehicle_id": { "type": "integer" } } } } } }, "responses": { "200": { "description": "Siker" } }
        },
        "delete": { "tags": ["Fuelings"], "summary": "Törlés", "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Siker" } } }
      },
	"/api/fuelings/vehicle/{id}": {
		"get": { "tags": ["Fuelings"], "summary": "Tankolások lekérdezése autó alapján", "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Siker" } } }
	},
...
    }
  }